move UI composition  details from the `EssentialFeediOS` to the `EssentialApp` module
